SERVERLESS-2327 Map environment into a context parameter
This translates our current environment variable based approach into passing a second "context" parameter to the function. This resembles the context objects we're introducing to all other languages.

Given the loosely typed nature of PHP, both using `Context` and `object` as the type for the parameters works. Since the user's IDE won't have the Context type in scope, it might bark at that potentially.